+++ doc/src/release_notes.rst
@@ -52,9 +52,9 @@ To see all issues & pull requests closed by this release see the
      :start-after: Version 4.1.0
      :end-before: .. rubric
 
-* pgr_maxWeightedMatching
+* pgr_maxWeightedMatch
 
-  .. include:: pgr_maxWeightedMatching.rst
+  .. include:: pgr_maxWeightedMatch.rst
      :start-after: Version 4.1.0
      :end-before: Description